# docs/openapi.yaml
openapi: 3.0.3
info:
  title: POMS API
  description: |
    **Print Operations Management System** — REST API for managing print job orders,
    staff tasks, inventory, billing, payments, and analytics.

    ## Authentication
    Every endpoint except `POST /auth/login` requires a Bearer JWT in the
    `Authorization` header.

    ## Error shape
    All errors return `{ "error": "<message>" }` with the appropriate HTTP status.

    ## Roles & permissions
    | Role | orders | tasks | billing | payments | inventory | machines | analytics | users |
    |------|--------|-------|---------|----------|-----------|----------|-----------|-------|
    | super_admin | CRUD | CRUD | CRUD | CRUD | CRUD | CRUD | R | CRUD |
    | admin | CRU | CRUD | R | CRU | CRU | CRU | R | CRU |
    | sub_admin | CRU | CRU | R | R | R | R | — | R |
    | designer | R | CRU (own) | — | — | R | — | — | — |
    | *_staff | R (own) | RU (own) | — | — | R | — | — | — |

  version: 1.0.0
  contact:
    email: capitalcurv8@gmail.com
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable parameters ────────────────────────────────────────────────────
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: 6655a1b2c3d4e5f6a7b8c9d0
    orderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        example: 6655a1b2c3d4e5f6a7b8c9d0

  # ─── Reusable responses ──────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
    BadRequest:
      description: Validation error or invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate unique field
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ─── Schemas ────────────────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth ────────────────────────────────────────────────────────────────────
...
    BillType:
      type: string
      enum: [raw, gst]
    BillLineItem:
      type: object
      properties:
        description:
          type: string
        qty:
          type: number
        rate:
          type: number
        amount:
          type: number
    Bill:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        type:
          $ref: '#/components/schemas/BillType'
        seriesNumber:
          type: string
          example: TAX-2024-001
        amount:
          type: number
        isProtected:
          type: boolean
          description: Raw bills are password-protected
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/BillLineItem'
        taxableAmount:
          type: number
          description: GST bills only
        cgst:
          type: number
          description: GST bills only — 9%
        sgst:
          type: number
          description: GST bills only — 9%
        totalAmount:
          type: number
          description: GST bills only — taxableAmount + cgst + sgst
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateBillRequest:
      type: object
      required: [orderId, type]
      properties:
        orderId:
          type: string
        type:
          $ref: '#/components/schemas/BillType'
    BillList:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: '#/components/schemas/Bill'
        total:
          type: integer

    # Inventory ───────────────────────────────────────────────────────────────
    MaterialCategory:
      type: string
      enum: [flex, paper, ink, acrylic, glass, other]
    MaterialUnit:
      type: string
      enum: [sqft, ml, sheets, kg, pcs]
    LedgerType:
      type: string
      enum: [DEDUCT, RESTOCK, REVERSAL, ADJUSTMENT]
    Material:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/MaterialCategory'
        unit:
          $ref: '#/components/schemas/MaterialUnit'
        stock:
          type: number
        threshold:
          type: number
          description: Low-stock alert fires when stock ≤ threshold
        costPerUnit:
          type: number
        supplier:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
              format: email
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    StockLedgerEntry:
      type: object
      properties:
        _id:
          type: string
        materialId:
          type: string
        orderId:
          type: string
        type:
          $ref: '#/components/schemas/LedgerType'
        qty:
          type: number
        balanceAfter:
          type: number
        note:
          type: string
        performedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateMaterialRequest:
      type: object
      required: [name, category, unit, threshold, costPerUnit]
      properties:
        name:
          type: string
          minLength: 2
        category:
          $ref: '#/components/schemas/MaterialCategory'
        unit:
          $ref: '#/components/schemas/MaterialUnit'
        stock:
          type: number
          minimum: 0
        threshold:
          type: number
          minimum: 0
        costPerUnit:
          type: number
          minimum: 0
        supplier:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
              format: email
    RestockRequest:
      type: object
      required: [qty]
      properties:
        qty:
          type: number
          minimum: 0
        note:
          type: string
    MaterialList:
      type: object
      properties:
        materials:
          type: array
          items:
            $ref: '#/components/schemas/Material'
        total:
          type: integer

    # Machines ────────────────────────────────────────────────────────────────
...
paths:
...
  /billing:
    post:
      tags: [Billing]
      summary: Generate a bill for an order
      description: |
        Two bill types:
        - **raw** — records the raw/unbranded cost; series `RAW-YYYY-NNN`; `isProtected: true`
        - **gst** — generates a tax invoice with CGST/SGST at 9% each; series `TAX-YYYY-NNN`

        Line items are auto-derived from `order.items`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBillRequest'
      responses:
        '201':
          description: Created bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Order not found
  /billing/order/{orderId}:
    get:
      tags: [Billing]
      summary: Get all bills for an order
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Bill list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /billing/{id}:
    get:
      tags: [Billing]
      summary: Get bill by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Inventory ───────────────────────────────────────────────────────────────
  /inventory/alerts:
    get:
      tags: [Inventory]
      summary: Get materials where stock ≤ threshold (low-stock alerts)
      responses:
        '200':
          description: Low-stock materials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /inventory:
    get:
      tags: [Inventory]
      summary: List all materials
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/MaterialCategory'
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Material list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Inventory]
      summary: Create a new material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaterialRequest'
      responses:
        '201':
          description: Created material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /inventory/{id}:
    get:
      tags: [Inventory]
      summary: Get material by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Inventory]
      summary: Update material metadata
      description: Does not adjust stock — use `/restock` for that.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                category:
                  $ref: '#/components/schemas/MaterialCategory'
                threshold:
                  type: number
                costPerUnit:
                  type: number
                supplier:
                  type: object
                  properties:
                    name:
                      type: string
                    phone:
                      type: string
                    email:
                      type: string
                      format: email
      responses:
        '200':
          description: Updated material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /inventory/{id}/restock:
    post:
      tags: [Inventory]
      summary: Increment material stock and log a RESTOCK ledger entry
      description: Runs inside a MongoDB session for atomicity.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestockRequest'
      responses:
        '200':
          description: Updated material and ledger entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  material:
                    $ref: '#/components/schemas/Material'
                  transaction:
                    $ref: '#/components/schemas/StockLedgerEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /inventory/{id}/ledger:
    get:
      tags: [Inventory]
      summary: Get full stock ledger for a material (newest first)
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Ledger entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockLedgerEntry'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Machines ────────────────────────────────────────────────────────────────
...